+++ internal/features/educators/models/case_manager.go
@@ -1,0 +1,10 @@
+package models
+
+import (
+	"seek/internal/features/students/models"
+)
+
+type CaseManager struct {
+	Educator
+	Caseload []models.Student
+}